moved question modal from common to components
